refactor: consolidate duplicate queue test mocks into shared test_helpers_test.go
Merge queueMockMatrixPort and otherMockMatrixPort into a single
testMockMatrixPort, eliminating duplication that would drift when
ports.MatrixPort changes.
